Refactor KingdomCards to not use classes
All Map<Class, Stack<Card>> has been refactored to a single Class. This might not be optimal, because we are not programming to the smallest possible interface anymore. It wouldn't be possible to use a different map, the question is if that is a use case that is probable in our case.